Add delete channel functionality
Added functions / methods to delete one channel from mongoDB and JSON local folder. I modified both mongo operations and infoYT. This means that we can directly cancel a channel by running the function in mongo_op but also, if we have an instance of InfoYT, we can use a method of that class. maybe the 2 functions are a little redundant. in InfoYT_streamlit I added an interactive functionality to delete the channel both from mongo and from JSON folder.