Update delete functionality
Updated the deletion of a channel functionality: left only the method in mongo_operations.py that will work on every channel from a dropdown menu. improved the visualization in the streamlit app and added a print when the channel is removed. we dropped the delete_channel() method in the InfoYT class because it is redundant and it limits the deletion to a previous initialized channel.