Added tags to video
Added methods to fetch tags for a video, save them in their table and link those tags table to the video table. Added also methods to query for tags or for tags in a channel or video. Correct the updating of a video without overwriting all the info as it did before.