Add channel and video managers
Still with the goal of modularity, file and respective classes for channel and video manager were added to the project. They are the high level classes which will be the point of contact for the actual request. They handle the calls to the low level methods defined in other classes.