Added playlist manager
Added the whole file with the new module playlist manager which is the high level class whose goal is to handle the requests for the playlist. added a new function in storage to delete a playlist. shifted the logic to define the default download directory from the downloader (which is a low level class) to the youtube client (higher level).