Finalized local syncing
Improved method in downloader to retrieve videos or shorts or lives by using yt-dlp. concluded the operations having the goal of syncing the local download folder with the current db present. the full process is divided into 4 steps and each method of the four step is implemented and a sync_library function handles and launch all the 4 steps one after the other.