Minor consistency checks
The main change comes because we want to be sure about consistency when we process a video. in particular some playlist have videos coming from other channels. if all videos are processed, then this means we add those other channels. we want to avoid this case to happen automatically without our consent. if a playlist has some video coming from other channels, we will skip those videos. Update also requirements for a newer version of yt-dlp.