Implementation of context manager protocol for client
Instead of manually close the connection of the database, let's implement the contextmanager protocol where you use the with ... and it automatically closes the connection once the with is done. this is also the best practice to handle such a client and also for any application where you are interacting with db.