Improvements on channel manager
added 2 functions that are just a calling for some functions at lower level. one retrieves the channel video list directly from youtube API, the other get the tags for the videos in the channel. then add a function which retrieves the videos currently locally donwloaded, via a new function created in storage, and compares it with those present in db but not yet downloaded.